// apps/web/src/components/UI/PremiumInput.d.ts
declare const PremiumInput: ({ label, placeholder, value, onChange, onBlur, onFocus, type, variant, size, disabled, error, helperText, maxLength, showCharCount, leftIcon, rightIcon, required, className, autoComplete, name, }: {
    label: any;
    placeholder: any;
    value?: string | undefined;
    onChange: any;
    onBlur: any;
    onFocus: any;
    type?: string | undefined;
    variant?: string | undefined;
    size?: string | undefined;
    disabled?: boolean | undefined;
    error: any;
    helperText: any;
    maxLength: any;
    showCharCount?: boolean | undefined;
    leftIcon: any;
    rightIcon: any;
    required?: boolean | undefined;
    className?: string | undefined;
    autoComplete: any;
    name: any;
}) => JSX.Element;
export default PremiumInput;
//# sourceMappingURL=PremiumInput.d.ts.map